#1253: Permetre firmar en el servidor tots els tipus de documents
Task-Url: https://github.com/GovernIB/ripea/issues/1253

+++ ripea-core/src/main/java/es/caib/ripea/core/helper/ExpedientHelper2.java
@@ -114,6 +114,11 @@ public class ExpedientHelper2 {
 		docsToDelete.addAll(documentRepository.findDeleted(expedientId));
 		
 		for (DocumentEntity docToDelete : docsToDelete) {
+			
+			if (CollectionUtils.isNotEmpty(docToDelete.getAnnexos())) {
+				throw new ValidationException("No está permitido esborrar documents procedents d'anotacions");
+			}
+			
 			documentHelper.deleteDefinitiu(docToDelete);
 		}
